# Optimize ExecutionStore#backendJobDescriptorKeyForNode. [BA-5767]