before
```
size -> 100000: 72.38088547222222 ms
size -> 200000: 118.50121208333333 ms
size -> 300000: 175.1667145 ms
size -> 400000: 209.58718594444443 ms
size -> 500000: 280.38145025 ms
size -> 600000: 337.16847691666663 ms
size -> 700000: 367.9980148888889 ms
size -> 800000: 426.0341651944443 ms
size -> 900000: 539.8858220277777 ms
size -> 1000000: 548.2584588333334 ms
```

after
```
size -> 100000: 0.020786249999999992 ms
size -> 200000: 0.016441083333333335 ms
size -> 300000: 0.01688913888888889 ms
size -> 400000: 0.015165277777777776 ms
size -> 500000: 0.01751997222222222 ms
size -> 600000: 0.006870222222222223 ms
size -> 700000: 0.056051750000000004 ms
size -> 800000: 0.018104638888888885 ms
size -> 900000: 0.009031027777777775 ms
size -> 1000000: 0.002292388888888889 ms
```