@cjllanwarne 2019-06-26
To save you from spelunking:
```
[0m[[0m[0minfo[0m] [0m[0m[31m- should convert a cloud sdk action *** FAILED *** (56 milliseconds)[0m[0m
[0m[[0m[0minfo[0m] [0m[0m[31m  "docker run google/[google/cloud-sdk:251.0.0-]slim" was not equal to "docker run google/[cloud-sdk:]slim" (ActionBuilderSpec.scala:50)
```